Refactor admin panel page to specific smaller chunks
